fix: incorrect project is created when template does not have App_Resources
In case the project does not have App_Resources dir, CLI tries to extract the content of the default template's App_Resources. However, as the structure of the default template has changed (to v2), CLI extracts the resources to incorrect location. This makes the newly created project unbuildable.
In order to fix this behavior and allow the default template to specify where App_Resources are located (via nsconfig.json file), change the logic in CLI to extract the default template to a temp location, read where the App_Resources is located and after that copy the content to the new project.