feat: remove requirement for Android Support Repository local installation
In the past there was a requirement to have a local installation of the Android Support Repository in order to use libraries from there. However, since some time, the new versions are not downloaded locally and they must be used from the Google repository.
Currently this is not possible in NativeScript as {N} CLI always passes the Support Library version to gradle based on the local installation.
Remove the requirement for local installation and never pass the version to gradle. This way the apps will use the value specified in the `build.gradle` and they can overwrite it in `app.gradle`.
Update nativescript-doctor to latest version, which allows building project without local Androud Support Repository