fix: unable to create project from scoped package template
In case you try to create application from scoped package, for example `@nativescript/some-package-name`, CLI fails with `npm install` error.
The problem is that CLI thinks everything after `@` is a version and sets the package name that has to be installed to empty string.
Fix the parsing and add tests