fix: passing --hmr must always set bundle option to "webpack"
In case `--hmr` is passed, CLI should set the `bundle` option to `webpack`. This works for normal command, but in case it is command that has its own specific `--options` (commands from extensions), CLI calls `validateOptions` again and resets the value set to `--bundle`.
In order to fix the behavior, set the `bundle` option to "webpack" directly in the setArgv method which is called every time.