refactor: remove OptionsBase and move logic to Options
As we do not have submodule anymore, remove the `OptionsBase` class and move the implementation to `Options` class.
Remove the `lib/common/options.ts` file and the `ICommonOptions` interface.
Replace the usage of `ICommonOptions` with `IOptions`.
Fix the unit tests of `lib/common/test/unit-tests/common-options` to use `Options` instead of `OptionsBase` and apply some changes in the passed options in the tests, so they can be valid now. Move the tests to `test/options.ts`