fix: tns preview warns incorrectly for exceeding file size limit
The check in `tns preview` relies on the zipped content instead of checking the actual file content. This way, based on what is the zipped content, we may end up with 30000 bytes length for 15 MB input files or for 100KBs input.
The check is not correct, it should rely on the input file size, as when extracted on device, their size is the actual limitation.