fix: tns debug ios command fails in some cases
`tns debug ios --start` and `tns debug ios --inspector` command fails in some cases as CLI is not able to get the debug port from the device logs. The problem is that ios-sim-portable starts the log process with child_process.spawn, but this does not start the actual process. Fix the case by waiting for the first message on the stdout of the log process (handled in ios-sim-portable). This way the actual logic that parses the logs will start after the process has really started.

Fix a case where we fail the connection to socket with `undefined` error - show correct error in such case.
Fix `tns device log` command of iOS devices/simulators - it must show full logs from device, but instead a predicate had been passed. Remove it from there and pass it only where required.