feat: set min required compile Android SDK to 28
As we are going to set min required Android Support Library to be 28 (in the modules), we need to set the min required compile SDK to 28 as well (as Android Support Library 28 relies on resources included in this version of the SDK and the project cannot be build without them). This way CLI will not allow building an application without installing the Android SDK 28 first.