fix: remove all .ts files when building in release
When building application in release mode, the `.ts` files of all plugins should be removed from the final application. Currently this is happening only for `tns-core-modules`. Do this for all the plugins.
Remove logic that deletes the `node_modules` directory of the `tns-core-modules` when it is copied to `platforms`. This is not required anymore and it breaks use-case when linking tns-core-modules.