fix: prepare node_modules only once
Whenever we prepare node_modules (i.e. moving them to platforms dir), CLI executes the operation twice:
1. When prepare the JS part of the modules.
2. When prepare the native part of the modules.
The first part is a must for cases when you work with cloud builds, so we cannot skip it, but we can safely skip the second one. This improves the initial prepare of `tns-template-master-detail-ng` with around 20 seconds.