feat: speed-up livesync with HMR
In case HMR is used, webpack reports to CLI the already prepared files, so there's no need for CLI to do any preparation, just sync the files. Try sending them as soon as possible and if we fail, go through the normal worklfow.

Ensure the code is called only when webpack had produced files for us and we are with HMR enabled.