feat(analytics): write analytics data to passed file
Write analytics data to a file passed as `--analyticsLogFile`. The file will contain timestamp when the action is executed and what is the data passed to analytics.
Also add some additional logging to the file, so when something fails, it will be easier for diagnostics.
In case `--analyticsLogFile` is not passed, information will not be tracked.