fix(analytics): undefined file is created on every command
When analytics are enabled, `undefined` file is created at the location where a command is executed. This is because the `$options.analyticsLoggingFile` is undefined, but we've set it as args to the spawned Analytics process. The value in the child process is the string "undefined", so that's why CLI always creates this file.
Pass the path to file only when it is passed by user.