fix: project properties are not tracked on every analytics hit
Currently projectType and isShared properties are tracked only when we send events to Google Analytics and when the caller sends the projectDir to the method. In fact we need to track them(based on the projectDir) for every hit in Google Analytics.
To achieve this, use the projectHelper, which tries to determine the projectDir. Track the projectType on every page and event send to Google Analytics.