fix: unable to use yarn on Windows
When you've set package manager to `yarn` on Windows, CLI is unable to create a project - it hangs during the operation. The problem is that getting the path to yarn cache returns result with `\n` at the end. This makes the `pacote` package hangs instead of just failing.
Fix this by trimming the result of `yarn cache dir` call.