fix: `tns update` does not work when yarn is set
When yarn is set as package manager, `tns update` command fails. The problem is that during command's execution, we try to remove `tns-core-modules-widgets` from the project. When `tns-core-modules-widgets` pacakage is not a dependency of the project (this is the default case as this package is dependency of `tns-core-modules` not a direct dependency of the project), calling `yarn remove` fails.
In the same case `npm` does not fail.
To resolve the issue, check if `tns-core-modules-widgets` are dependency of the project and uninstall this package only in such case.