fix: handle `previewAppLiveSyncError` event in CLI
Currently the `previewAppLiveSyncError` event is handled only when CLI is used as a library. Add handler when CLI is used from command line and use the same workflow in both cases.
Attach to the event every time when a new operation is started. Currently, once LiveSync is stopped, we've removed all handlers and we were never attaching them again. Now we rely on the isInitialized property, so we'll attach the event handler whenever it is needed.