fix: execute only attach request in order to enable the ios debugging
We don't need the AttachAvailability response as now we don't need different actions based on the response. We just need to get the debug port which is printed by the AttachRequest and the Runtime is now accepting connections regardless of its state. This also allows us to remove the availability calls from the runtime.