fix: error is raised when DevTools are closed while sending message
In case you open Chrome DevTools and close them while the iOS runtime still sends messages, an error is raised in CLI:
```
Error: WebSocket is not open: readyState 2 (CLOSING)
```

The problem is that we are using Transform stream between the communication and close event of the socket is not raised until the stream flushes its current state.
Fix it by skipping the write operation in case the socket is not in OPEN state.