fix: error may be raised during restart of app on ios Sim
On slower mac machines, an error may be raised during restart of app on iOS Simulator. The problem is that we stop the application and try to start it after that. However, the stop method may be resolved before the application is actually killed. Trying to start it after that leads to undefined behavior and errors in some cases.
To fix this, update ios-sim-portable to 4.0.8 where a possible fix is applied.