feat: skip postinstall steps in case CLI is not installed globally
In case the CLI package is not installed globally, probably it is installed as a dependency of a project or as a dependency of other dependency.
In this case the postinstall actions have no meaning for the user, as probably CLI will be used as a library.
Skip all of the postinstall actions in such case. Remove the execution code of the old `post-install` command and move it to `post-install-cli` command.
Add unit tests. Remove a property from staticConfig - it has no meaning to be there, just place it in the postinstall command directly.