chore: remove dynamicHelpService and dynamicHelpProvider
Remove `dynamicHelpService` and `dynamicHelpProvider` - they were used when the code between AppBuilder CLI and NativeScript CLI was shared. The idea of the service was to get the help content of ScreenBuilder commands. Now we do not need such logic, so just remove it and clean the code from the unused interfaces.