fix: high cpu usage during livesync
During LiveSync operations (i.e. `tns run ...`), CLI may keep the CPU of the machine at a constantly high level even when it is not livesyncing. This is caused by the device detection changes that we are starting, which also checks for new emulator images.
It turnes out the command `avdmanager list avds` takes really high usage of CPU on some machines when there are available AVD images.
To fix this behavior split the current interval on two different ones - one will check for devices, which is run on 200ms and a second one which will check for emulator images, which will run once per minute.
Fix the `startDeviceDetectionInterval` method - it should just start the device detection instead of waiting for the first execution to receive results. This breaks the current behavior, so execute a single synchronous device detection before starting the interval - this way we keep the current behavior.