fix: add require of lodash in helpers
Currently the postinstall.js requires helpers and tries to execute some methods of it. However, CLI is not bootstraped and the global variable `_` is not set, so the action fails. Fix the code by adding require to `lodash` in the current file.
The other option of fix is to remove the `helpers` require from postinstall script to `post-install-cli` command. However, the current check is if we should execute the postinstall actions - if we move the check to the command, the commands service will still execute the check and prompt the user to allow tracking, which is not a desired behavior.