fix: incorrect filtering of Android logs by tag
CLI filters Android logs by PID and by tags. However, in case the PID matches, but we do not find the tag, we may still output the line as we check if it contains the tag anywhere in the content.
So, when searching for `JS` tag, we may output the line:
```
02-01 13:29:23.990  8141  8141 I art     :   at java.lang.Object com.tns.Runtime.callJSMethod(java.lang.Object, java.lang.String, java.lang.Class, boolean, long, java.lang.Object[]) (Runtime.java:987)
```

Remove this incorrect behavior and rely on the tag filtering.