chore: fix travis build of next/rc/release-patch versions
Currently the Travis build of `next`, `rc` and `release-patch` versions takes way too long as the packaging and execution of unit tests is done twice - first time from `grunt pack` execution inside the script and the second time from the npm deploy provider.
To fix this introduce new Grunt task - `travisPack` - it checks if the build is from PR and in case it is, the grunt pack task is skipped. This way the `grunt travisPack` command in the `.travis.yml` script section will be skipped in case we are not in PR build, but the one from npm deploy step will still be executed.

Also delete some unused and not needed logic.