chore(travis): set correct version in package.json
The environment variable `RELEASE_BUILD` is used in Jenkins job, when CLI has to be build with strict official version. It's been deleted in a previous commit, which broke this logic - get it back and add comment why we need it.
Also ensure Travis will not push strict version in `npm` by calling `set_package_version` in the travisPack grunt task. In case it is not called, the Travis job, that is executed on push in master/release/release-patch branch, will execute `npm publish` only (as the `travisPack` will not do anything). The prepack script will trigger `grunt prepare` logic, which will not change the version in package.json, so it will remain `5.1.1` for example.
We cannot change the version in the prepack script (i.e. to call `set_package_version` in the `grunt prepare` task, as npm has problem with this approach - it produces `.tgz` of the built package with correct version, however it still tries to publish the version that was originally in the package.json.