chore(travis): remove duplicate npm deploy providers
Currently we have npm deploy provider for each branch that we want to publish on push. By design Travis will not call the deployment providers on PR builds, so they are executed only on push.
Our current travis build is triggered only on push in three branches - `master`, `release` and `release-patch`. So there's no need to have the same logic about deploying in npm three times as it cannot be executed for any other branch at the moment.
So keep a single deploy mechanism and allow it for all branches.