fix: set correct versions in plugin's build.gradle
With Gradle 3.3.0 (used in tns-android 5.2.0-... ) dependencies checking is stricter and there are some problems in the resolution of the supportLibrary version when building plugins. Issue is raised when the plugin has any of the following:
* another compile dependency which is built with `compile` of android support library version that is not the same as the one we are using
* plugin defines in its gradle new version of android support library, that is different from the one specified in our build.gradle.

To resolve this, set all versions of `com.android.support` libraries to the passed supportVersion or 28.0.0
Also update other tooling versions to latest versions.