chore: skip tns_modules cleanup from CLI
In case `--bundle` is passed CLI tries to clean the `tns_modules` directory produced from previous prepare operations. However, this is not required in latest versions as webpack cleans the directory on its own.
The current clean was required for Webpack before 0.14.0, particularly the one used with 3.x versions of NativeScript.
Delete unit test that is no longer relevant, as it tests functionality that is now handled by `nativescript-dev-webpack` plugin.