feat: detect and report short imports used in application
In NativeScript 5.2.0 we've deprecated support for short imports. Add ability in CLI to check if there are such imports in the application's code and report them to users. The logic will be executed whenever application is prepared for build/livesync or when `tns doctor` is executed inside project dir.
Create class with static variables for ProjectTypes. These values must not be changed as they are used in Analytics, so changing them will break the current reports.