fix: prompts to install javac should not be shown on macOS
macOS has handle when you try to execute java or javac to prompt to install it. However, we do not want to bother the user with such prompts. Currently we are executing `javac -version` to determine the version. Instead of this, check if there's `javac` in the PATH before executing the command. In case there's no javac in the path, do not execute `javac -version`, so we'll not prompt the user to install it.