fix: do not start emulator detection interval
The emulator detection interval is not required for CLI and in fact it just uses CPU and memory. Also, it causes some issues when you do not hava JAVA installed.
So remove start of emulator detection interval and allow it to be started from the API, when CLI is used as a library.