fix: javac detection does not work as required by Gradle
The current `javac` detection tries to get the `javac` path from JAVA_HOME and to verify it is correct. In case it fails somewhere, it fallbacks to check the `javac` from PATH. So, from users' perspective it seems javac is setup correctly.
However, Gradle works in slightly different manner - it checks if you have JAVA_HOME and if so - verifies `java` from it. In case there's no `java` in the `$JAVA_HOME/bin/java`, an error is thrown.
Make our check for `javac` in the same way - in case you have set JAVA_HOME, we'll verify `javac` from there. This is exactly the same as the one we had prior 1.9.0 version, so this is just a regression fix.
Add unit tests to cover this scenario.