feat: improve cleanup service
Currently it is possible just to add new actions to be executed in the cleanup. However, in some cases, we want to add action and remove it after that.
Also, we need to delete some lock files when CLI is not using them anymore. So, introduce new API in the cleanup service that allow all of these actions.