feat: allow setting timeout when using childProcess
In Node.js you can set timeout option to child processes started with `exec`. However, there's no support for this when using `spawn`.
Add support for passing timeout in our abstraction of childProcess, when using `spawnFromEvent` method. When this method is used, it blocks the operation until the process finishes its execution. So it is appropriate to pass timeout and if it passes, send `SIGTERM` signal to the child process. This is the same signal used by Node.js when the timeout is hit in `child_process.exec`.