fix: fix the order of applying Podfiles
Currently when CLI adds plugin's Podfile, it places its content at the top of the newly generated Podfile. Same happens for the post_install hook of the plugin. The last Podfile that is applied, is the one from App_Resources/iOS, but its content is also placed at the top of the merged Podfile and its post_install action (in case it has such) is executed first.
The idea of having Podfile in App_Resources/iOS was to be able to overwrite some settings from plugins Podfiles, for example SWIFT_VERSION.
To handle this, append each Podfile at the end of the merged Podfile. Same is valid for the post_install hook coming from a Podfile - it will be appended at end of the merged post_install function.