feat: introduce initializeService
Introduce initializeService which can be used to setup some specifics of CLI (currently the logger) and to print all required warnings. At the moment, whenever we want to introduce new warnings, we need to introduce new methods that we call from CLI and when it is used as a library.
With the new solution all such warnings can be placed in the initialize method and it will automatically show those warnings.