fix: ensure analytics and cleanup processes are not leaking
When CLI dies, the `analyticsBroker` and `cleanupProcess` processes receive disconnect event. At this point they have to finish their work and die silently after that.
However, this does not happen for `analyticsBroker` process as in the `disconnect` handler we call `process.disconnect` method. It fails, as the process is already disconnected. So the process hangs in undefined state and it leaks. The stareted cleanup process from it also leaks.
To fix this remove the disconnect call from the handler.
Also ensure injector is disposed before calling process.exit in the detached process's handler.