fix: logger changes broke trace and some hooks
`--log trace` is broken as we are tring to iterate over object created with `Object.create(null)`. It uses the prototype of null, so it does not have `hasOwnProperty` method. Filter such objects from our checks in logger.
Also nativescript-plugin-firebase uses logger.out, which has been deleted, get it back for backwards compatibility and delete it in 6.0.0 release.